[BE] refactor: Accommodation package 구조 개선
- web 안에 controller 와 model 이 같이 있었는데, 이중 model 을 web-model 패키지를 만들어 이동시켰습니다
