+++ BE/airbnb/src/main/java/com/codesquad/airbnb/accmmodation/business/AccommodationService.java
@@ -3,9 +3,9 @@ package com.codesquad.airbnb.accmmodation.business;
 import com.codesquad.airbnb.accmmodation.data.Accommodation;
 import com.codesquad.airbnb.accmmodation.data.AccommodationRepository;
 import com.codesquad.airbnb.accmmodation.data.Price;
-import com.codesquad.airbnb.accmmodation.web.AccommodationQuery;
-import com.codesquad.airbnb.accmmodation.web.AccommodationView;
-import com.codesquad.airbnb.accmmodation.web.DetailAccommodationView;
+import com.codesquad.airbnb.accmmodation.web.model.AccommodationQuery;
+import com.codesquad.airbnb.accmmodation.web.model.AccommodationView;
+import com.codesquad.airbnb.accmmodation.web.model.DetailAccommodationView;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
