+++ BE/airbnb/src/main/java/com/codesquad/airbnb/accmmodation/web/model/AccommodationView.java
@@ -1,30 +1,31 @@
-package com.codesquad.airbnb.accmmodation.web;
+package com.codesquad.airbnb.accmmodation.web.model;
 
 import com.codesquad.airbnb.accmmodation.data.Accommodation;
-import com.codesquad.airbnb.accmmodation.data.Coordinate;
 import com.codesquad.airbnb.accmmodation.data.Image;
+import com.codesquad.airbnb.accmmodation.data.Price;
 import com.codesquad.airbnb.accmmodation.data.type.AccommodationType;
+import com.codesquad.airbnb.accmmodation.data.type.ImageType;
 import java.util.List;
 import lombok.Builder;
 import lombok.Getter;
 
 @Getter
-public class DetailAccommodationView {
+public class AccommodationView {
 
   private final Long id;
   private final String name;
   private final AccommodationType type;
   private final String location;
-  private final Coordinate coordinate;
+  private final Price price;
   private final List<Image> images;
 
   @Builder
-  public DetailAccommodationView(Accommodation accommodation) {
+  public AccommodationView(Accommodation accommodation) {
     this.id = accommodation.getId();
     this.name = accommodation.getName();
     this.type = accommodation.getType();
     this.location = accommodation.getLocation();
-    this.coordinate = accommodation.getCoordinate();
-    this.images = accommodation.getImages();
+    this.price = accommodation.getPrice();
+    this.images = accommodation.getImages(ImageType.MAIN);
   }
 }
